+++ src/quacc/settings.py
@@ -277,7 +277,7 @@ class QuaccSettings(BaseSettings):
 
     @field_validator("PRIMARY_STORE")
     def generate_store(cls, v):
-        return MontyDecoder().decode(v) if v else None
+        return MontyDecoder().decode(v) if isinstance(v, str) else v
 
     model_config = SettingsConfigDict(env_prefix="quacc_")